// migrations/snapshots/f6476841b38b63a24990558a0a0b2daf2afc8ae0185d672b05ebe282971725b4/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Area: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Buyer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly areaId: CodecTypes['pg/int4@1']['output'];
      readonly length: CodecTypes['pg/int4@1']['output'];
      readonly width: CodecTypes['pg/int4@1']['output'];
      readonly budget: CodecTypes['pg/int4@1']['output'];
      readonly status: 'NOTPURCHASED' | 'PURCHASED' | 'CANCELLED';
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Client: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly address: CodecTypes['pg/text@1']['output'];
      readonly mobileNumber: CodecTypes['pg/text@1']['output'];
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Colony: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Employee: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly phoneNumber: CodecTypes['pg/text@1']['output'];
    };
    readonly Flat: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly floorId: CodecTypes['pg/int4@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly flatNumber: CodecTypes['pg/text@1']['output'];
      readonly status: 'AVAILABLE' | 'SOLD';
    };
    readonly Floor: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly wingId: CodecTypes['pg/int4@1']['output'];
      readonly numberOfFlats: CodecTypes['pg/int4@1']['output'];
    };
    readonly FollowUp: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly employeeId: CodecTypes['pg/int4@1']['output'];
      readonly dateTime: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly notes: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'PENDING' | 'DONE' | 'POSTPONED' | 'CANCELLED';
      readonly purpose: 'GENERAL' | 'MARKETING' | 'BUYER' | 'SELLER';
    };
    readonly Marketing: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly employeeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly colonyId: CodecTypes['pg/int4@1']['output'];
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'NONE' | 'CALLED_NOT_VISITED' | 'VISITED' | 'NOT_INTERESTED';
      readonly priority: 'ONE_STAR' | 'TWO_STAR' | 'THREE_STAR' | 'FOUR_STAR' | 'FIVE_STAR';
    };
    readonly Seller: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly areaId: CodecTypes['pg/int4@1']['output'];
      readonly length: CodecTypes['pg/int4@1']['output'];
      readonly width: CodecTypes['pg/int4@1']['output'];
      readonly demand: CodecTypes['pg/int4@1']['output'];
      readonly status: 'UNSOLD' | 'SOLD';
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Wing: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly colonyId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly numberOfFloors: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Area: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Buyer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clientId: CodecTypes['pg/int4@1']['input'];
      readonly areaId: CodecTypes['pg/int4@1']['input'];
      readonly length: CodecTypes['pg/int4@1']['input'];
      readonly width: CodecTypes['pg/int4@1']['input'];
      readonly budget: CodecTypes['pg/int4@1']['input'];
      readonly status: 'NOTPURCHASED' | 'PURCHASED' | 'CANCELLED';
      readonly remarks: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Client: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly address: CodecTypes['pg/text@1']['input'];
      readonly mobileNumber: CodecTypes['pg/text@1']['input'];
      readonly remarks: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Colony: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Employee: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly phoneNumber: CodecTypes['pg/text@1']['input'];
    };
    readonly Flat: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly floorId: CodecTypes['pg/int4@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly flatNumber: CodecTypes['pg/text@1']['input'];
      readonly status: 'AVAILABLE' | 'SOLD';
    };
    readonly Floor: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly wingId: CodecTypes['pg/int4@1']['input'];
      readonly numberOfFlats: CodecTypes['pg/int4@1']['input'];
    };
    readonly FollowUp: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clientId: CodecTypes['pg/int4@1']['input'];
      readonly employeeId: CodecTypes['pg/int4@1']['input'];
      readonly dateTime: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly notes: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'PENDING' | 'DONE' | 'POSTPONED' | 'CANCELLED';
      readonly purpose: 'GENERAL' | 'MARKETING' | 'BUYER' | 'SELLER';
    };
    readonly Marketing: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clientId: CodecTypes['pg/int4@1']['input'];
      readonly employeeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly colonyId: CodecTypes['pg/int4@1']['input'];
      readonly remarks: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'NONE' | 'CALLED_NOT_VISITED' | 'VISITED' | 'NOT_INTERESTED';
      readonly priority: 'ONE_STAR' | 'TWO_STAR' | 'THREE_STAR' | 'FOUR_STAR' | 'FIVE_STAR';
    };
    readonly Seller: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clientId: CodecTypes['pg/int4@1']['input'];
      readonly areaId: CodecTypes['pg/int4@1']['input'];
      readonly length: CodecTypes['pg/int4@1']['input'];
      readonly width: CodecTypes['pg/int4@1']['input'];
      readonly demand: CodecTypes['pg/int4@1']['input'];
      readonly status: 'UNSOLD' | 'SOLD';
      readonly remarks: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Wing: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly colonyId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly numberOfFloors: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly area: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly buyer: {
      readonly areaId: CodecTypes['pg/int4@1']['output'];
      readonly budget: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly length: CodecTypes['pg/int4@1']['output'];
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'NOTPURCHASED' | 'PURCHASED' | 'CANCELLED';
      readonly width: CodecTypes['pg/int4@1']['output'];
    };
    readonly client: {
      readonly address: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly mobileNumber: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly colony: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly employee: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly phoneNumber: CodecTypes['pg/text@1']['output'];
    };
    readonly flat: {
      readonly flatNumber: CodecTypes['pg/text@1']['output'];
      readonly floorId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: 'AVAILABLE' | 'SOLD';
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly floor: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly numberOfFlats: CodecTypes['pg/int4@1']['output'];
      readonly wingId: CodecTypes['pg/int4@1']['output'];
    };
    readonly followUp: {
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly dateTime: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly employeeId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly notes: CodecTypes['pg/text@1']['output'] | null;
      readonly purpose: 'GENERAL' | 'MARKETING' | 'BUYER' | 'SELLER';
      readonly status: 'PENDING' | 'DONE' | 'POSTPONED' | 'CANCELLED';
    };
    readonly marketing: {
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly colonyId: CodecTypes['pg/int4@1']['output'];
      readonly employeeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: 'ONE_STAR' | 'TWO_STAR' | 'THREE_STAR' | 'FOUR_STAR' | 'FIVE_STAR';
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'NONE' | 'CALLED_NOT_VISITED' | 'VISITED' | 'NOT_INTERESTED';
    };
    readonly seller: {
      readonly areaId: CodecTypes['pg/int4@1']['output'];
      readonly clientId: CodecTypes['pg/int4@1']['output'];
      readonly demand: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly length: CodecTypes['pg/int4@1']['output'];
      readonly remarks: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'UNSOLD' | 'SOLD';
      readonly width: CodecTypes['pg/int4@1']['output'];
    };
    readonly wing: {
      readonly colonyId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly numberOfFloors: CodecTypes['pg/int4@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;